new interfaces for 1D and 2D splines, ND splines

+++ Detectors/TPC/reconstruction/include/TPCReconstruction/TPCFastTransformHelperO2.h
@@ -74,7 +74,7 @@ class TPCFastTransformHelperO2
   /// initialization
   void init();
   /// get space charge correction in internal TPCFastTransform coordinates su,sv->dx,du,dv
-  int getSpaceChargeCorrection(int slice, int row, float su, float sv, float& dx, float& du, float& dv);
+  int getSpaceChargeCorrection(int slice, int row, double su, double sv, double& dx, double& du, double& dv);
 
   static TPCFastTransformHelperO2* sInstance;                                                  ///< singleton instance
   bool mIsInitialized = 0;                                                                     ///< initialization flag
